Relocate `FrameCountPlugin` to `bevy_diagnostic`
