--- crates/bevy_core/src/serde.rs
@@ -1,54 +1,0 @@
-use core::{
-    any,
-    fmt::{self, Formatter},
-};
-
-use serde::{
-    de::{Error, Visitor},
-    Deserialize, Deserializer, Serialize, Serializer,
-};
-
-use super::FrameCount;
-
-// Manually implementing serialize/deserialize allows us to use a more compact representation as simple integers
-impl Serialize for FrameCount {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        serializer.serialize_u32(self.0)
-    }
-}
-
-impl<'de> Deserialize<'de> for FrameCount {
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        deserializer.deserialize_u32(FrameVisitor)
-    }
-}
-
-struct FrameVisitor;
-
-impl<'de> Visitor<'de> for FrameVisitor {
-    type Value = FrameCount;
-
-    fn expecting(&self, formatter: &mut Formatter) -> fmt::Result {
-        formatter.write_str(any::type_name::<FrameCount>())
-    }
-
-    fn visit_u32<E>(self, v: u32) -> Result<Self::Value, E>
-    where
-        E: Error,
-    {
-        Ok(FrameCount(v))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use serde_test::{assert_tokens, Token};
-
-    #[test]
-    fn test_serde_frame_count() {
-        let frame_count = FrameCount(100);
-        assert_tokens(&frame_count, &[Token::U32(100)]);
-    }
-}
