+++ crates/bevy_core_pipeline/src/taa/mod.rs
@@ -8,7 +8,7 @@ use crate::{
 };
 use bevy_app::{App, Plugin};
 use bevy_asset::{load_internal_asset, Handle};
-use bevy_core::FrameCount;
+use bevy_diagnostic::FrameCount;
 use bevy_ecs::{
     prelude::{require, Bundle, Component, Entity, ReflectComponent},
     query::{QueryItem, With},
