+++ crates/bevy_diagnostic/Cargo.toml
@@ -12,6 +12,7 @@ keywords = ["bevy"]
 # Disables diagnostics that are unsupported when Bevy is dynamically linked
 dynamic_linking = []
 sysinfo_plugin = ["sysinfo"]
+serialize = ["dep:serde"]
 
 [dependencies]
 # bevy
@@ -23,6 +24,7 @@ bevy_utils = { path = "../bevy_utils", version = "0.15.0-dev" }
 bevy_tasks = { path = "../bevy_tasks", version = "0.15.0-dev" }
 
 const-fnv1a-hash = "1.1.0"
+serde = { version = "1.0", optional = true }
 
 # macOS
 [target.'cfg(all(target_os="macos"))'.dependencies]
@@ -37,6 +39,9 @@ sysinfo = { version = "0.33.0", optional = true, default-features = false, featu
 sysinfo = { version = "0.33.0", optional = true, default-features = false, features = [
   "system",
 ] }
+
+[dev-dependencies]
+serde_test = "1.0"
 
 [lints]
 workspace = true
