+++ crates/bevy_internal/src/default_plugins.rs
@@ -6,7 +6,7 @@ plugin_group! {
         bevy_app:::PanicHandlerPlugin,
         bevy_log:::LogPlugin,
         bevy_core:::TaskPoolPlugin,
-        bevy_core:::FrameCountPlugin,
+        bevy_diagnostic:::FrameCountPlugin,
         bevy_time:::TimePlugin,
         bevy_transform:::TransformPlugin,
         bevy_hierarchy:::HierarchyPlugin,
@@ -107,7 +107,7 @@ plugin_group! {
     /// This plugin group will add the minimal plugins for a *Bevy* application:
     pub struct MinimalPlugins {
         bevy_core:::TaskPoolPlugin,
-        bevy_core:::FrameCountPlugin,
+        bevy_diagnostic:::FrameCountPlugin,
         bevy_time:::TimePlugin,
         bevy_app:::ScheduleRunnerPlugin,
         #[cfg(feature = "bevy_ci_testing")]
